objtype: prototype the nine objGetNearestTypeTo callers the 21-TU roster missed
The -requireprotos census found nine more TUs calling through the
implicit int declaration: engine/46, engine/63, 218, 251, 278, 285,
389, 623, 649. Same treatment; 251/285 u32 receivers spell the
conversion explicitly. All nine objects neutral in every linked
section, relocation and symbol.

+++ src/dlls/objects/251/251.c
@@ -11,6 +11,7 @@
 #include "main/frame_timing.h"
 #include "main/gamebits_api.h"
 #include "main/objtexture.h"
+#include "main/objtype.h"
 #include "main/vecmath_distance_api.h"
 
 #define PRESSURESWITCHFB_PARTFX_ID                      0x7C3
@@ -179,7 +180,7 @@ void PressureSwitchFB_update(GameObject* obj) {
             state->contactTimer = 0;
         }
         searchRadius = PRESSURESWITCHFB_TARGET_SEARCH_RADIUS;
-        nearestTarget = objGetNearestTypeTo(PRESSURESWITCHFB_TARGET_OBJECT_GROUP, obj, &searchRadius);
+        nearestTarget = (u32)objGetNearestTypeTo(PRESSURESWITCHFB_TARGET_OBJECT_GROUP, obj, &searchRadius);
         if (nearestTarget != 0) {
             state->contactTimer = PRESSURESWITCHFB_CONTACT_FRAMES;
         }
